Progress bar and args parser

+++ test/test.js
@@ -2,18 +2,21 @@ const exec = require('util').promisify(require('child_process').exec);
 const fs = require('fs');
 const assert = require('assert');
 const mkgif = require('../mkgif.js');
+const DELETE_TEMP_FILE = false;
 
 function test(inputArgs) {
   return new Promise(async function(resolve, reject) {
-    const outputFileName = 'test/temp-water.gif';
-    const inputFileName = 'test/water.mp4';
-    const {stdout, stderr} = await mkgif([inputFileName, outputFileName]);
+    const inputFileName = 'test/test.m4v';
+    const outputFileName = 'test/temp-test.gif';
+    await mkgif({ _: [ inputFileName, outputFileName ] });
     fs.stat(outputFileName, function (err, stats) {
       assert(stats.size > 100000); // must be greater than 100 thousand
-      fs.unlink(outputFileName, (err) => { // delete temp output file
-        if (err) reject(err);
-        resolve({stdout, stderr});
-      });
+      if (DELETE_TEMP_FILE) {
+        fs.unlink(outputFileName, (err) => { // delete temp output file
+          if (err) reject(err);
+          resolve(stats.size);
+        });
+      } else resolve(stats.size);
     });
   });
 };
